Moved search logic to querysets
