+++ hexa/catalog/views.py
@@ -8,9 +8,6 @@ from django.shortcuts import get_object_or_404, redirect, render
 from django.utils.translation import ugettext_lazy as _
 from django.views.decorators.http import require_http_methods
 
-from hexa.core.string import tokenize
-from hexa.user_management.models import User
-
 from .datagrids import DatasourceGrid
 from .models import Datasource, Index
 from .queue import datasource_sync_queue
@@ -39,46 +36,16 @@ def index(request: HttpRequest) -> HttpResponse:
     )
 
 
-def _build_search_query(user: User, raw_query: str):
-    tokens = tokenize(raw_query)
-
-    # filters
-    types = [t.value[5:] for t in tokens if t.value.startswith("type:")]
-    datasources = []
-    for t in tokens:
-        if t.value.startswith("datasource:"):
-            try:
-                datasources.append(uuid.UUID(t.value[11:]))
-            except ValueError:
-                continue
-
-    # query
-    results = (
-        Index.objects
-        # only select resources viewable by user
-        .filter_for_user(user)
-        # rank resource based on exact words/related words
-        .search(tokens)
-        # filter by resources type
-        .filter_for_types(types)
-        # filter by datasources
-        .filter_for_datasources(datasources)
-        # exclude s3keep, artifact of s3content mngt
-        .exclude(external_name=".s3keep")
-    )
-    return results
-
-
 def quick_search(request: HttpRequest) -> HttpResponse:
     query = request.GET.get("query", "")
-    results = _build_search_query(request.user, query)[:10]
-    print([result.to_dict() for result in results])
+    results = Index.objects.filter_for_user(request.user).search(query)[:10]
+
     return JsonResponse({"results": [result.to_dict() for result in results]})
 
 
 def search(request: HttpRequest) -> HttpResponse:
     query = request.GET.get("query", "")
-    results = _build_search_query(request.user, query)[:100]
+    results = Index.objects.filter_for_user(request.user).search(query)[:100]
 
     types, datasources = [], []
     for ct in ContentType.objects.filter(app_label__startswith="connector_"):
@@ -97,7 +64,6 @@ def search(request: HttpRequest) -> HttpResponse:
         if hasattr(model, "searchable"):
             types.append(f"{ct.app_label[10:]}_{ct.model}")
 
-    print([result.to_dict() for result in results])
     return render(
         request,
         "catalog/search.html",
